@@ -1,1 +1,1 @@
-Search.setIndex({"docnames": ["LT", "LT_Fit", "index", "install", "introduction", "src/lt_box", "src/lt_datafile", "src/lt_fit_gen_fit", "src/lt_fit_linear_fit", "src/lt_fit_parameters", "src/lt_mca", "src/lt_parameterfile", "src/lt_pdatafile", "src/lt_plotting"], "filenames": ["LT.rst", "LT_Fit.rst", "index.rst", "install.rst", "introduction.rst", "src/lt_box.rst", "src/lt_datafile.rst", "src/lt_fit_gen_fit.rst", "src/lt_fit_linear_fit.rst", "src/lt_fit_parameters.rst", "src/lt_mca.rst", "src/lt_parameterfile.rst", "src/lt_pdatafile.rst", "src/lt_plotting.rst"], "titles": ["The LT Package", "The LT_Fit Package", "The Main Packages", "Installation", "Introduction/Tutorial", "box", "datafile", "gen_fit", "linear_fit", "parameters", "MCA", "parameterfile", "pdatafile", "plotting"], "terms": {"A": [0, 5], "collect": 0, "python": [0, 3, 11], "softwar": [0, 3, 10], "tool": [0, 1, 5, 7, 8, 9], "data": [0, 4, 5, 6, 7, 8, 11, 12, 13], "analysi": [0, 4, 5], "present": [0, 5, 6, 12], "experiment": [0, 5, 7, 13], "physic": [0, 13], "ar": [0, 3, 5, 6, 7, 8, 9, 11, 12, 13], "base": [0, 5, 7, 13], "numpi": [0, 5, 6, 7, 8, 13], "scipi": [0, 7], "matplotlib": [0, 5, 7, 13], "recommend": [0, 3], "interact": [0, 13], "shell": 0, "i": [0, 5, 6, 7, 8, 9, 11, 12, 13], "ipython": [0, 13], "content": [0, 1, 2, 5, 6, 12, 13], "box": [0, 2, 7], "get_fil": [0, 5], "get_data": [0, 5, 6, 11, 12, 13], "get_spectrum": [0, 5], "histo": [0, 5], "histo2d": [0, 5], "histogram": 0, "oper": [0, 9], "datafil": [0, 2, 5, 11, 12, 13], "dfile": [0, 6, 12, 13], "parameterfil": [0, 2, 12], "pfile": [0, 11, 12], "pdatafil": [0, 2, 5], "pdfile": [0, 12], "plot": [0, 2, 5, 7, 8], "plot_exp": [0, 5, 7, 13], "plot_lin": [0, 5, 7, 13], "plot_splin": [0, 5, 13], "datafile_plot_exp": [0, 5, 13], "datafile_plot_theori": [0, 5, 13], "datafile_spline_plot_theori": [0, 5, 13], "dplot_exp": [0, 5, 13], "dplot_lin": [0, 5, 13], "dplot_splin": [0, 5, 13], "mca": [0, 2, 5], "fit": [1, 5, 7, 8, 9], "labtools3": [1, 3], "paramet": [1, 2, 5, 6, 7, 8, 11, 12], "linear_fit": [1, 2], "linefit": [1, 5, 8], "polyfit": [1, 5, 8], "gen_linfit": [1, 5, 8], "gen_fit": [1, 2, 5], "genfit": [1, 5, 7], "lt": [2, 4, 5, 6, 7, 10, 11, 12, 13], "lt_fit": [2, 4, 5, 7, 8, 9], "instal": [2, 6], "us": [2, 4, 5, 6, 7, 8, 9, 10, 11, 13], "anaconda": 2, "introduct": 2, "tutori": [2, 3], "index": [2, 6], "modul": [2, 4, 5, 7, 8, 10, 11, 13], "search": 2, "page": 2, "These": [3, 5, 9], "instruct": 3, "assum": [3, 5, 6, 10, 11, 13], "you": [3, 5, 6, 7, 8, 10, 13], "have": [3, 5, 10, 11], "all": [3, 5, 6, 8, 11, 12, 13], "necessari": [3, 4, 5], "requir": 3, "run": 3, "import": [3, 5, 7, 8, 9, 10, 12, 13], "onli": [3, 5, 6, 7, 9, 13], "after": 3, "ani": [3, 11, 13], "other": [3, 6, 9, 11], "distribut": 3, "want": [3, 5, 6, 8], "There": [3, 13], "sever": 3, "method": [3, 5], "avail": [3, 6], "download": 3, "script": 3, "install_labtool": 3, "py": 3, "from": [3, 5, 11, 12, 13], "command": [3, 13], "line": [3, 5, 6, 7, 8, 11, 12, 13], "pip": 3, "make": [3, 5, 6, 7, 8], "sure": [3, 5, 6], "administr": 3, "provoledg": [], "window": [3, 5], "maco": 3, "linux": 3, "might": [3, 7], "need": [3, 5, 6, 7, 13], "sudo": 3, "start": [3, 5, 6, 11, 12], "spyder": 3, "file": [3, 5, 6, 10, 11, 12], "explor": 3, "go": 3, "directori": 3, "folder": 3, "where": [3, 4, 5, 6, 8], "extract": [], "labtool": 4, "figur": 3, "1": [3, 5, 6, 7, 9, 12, 13], "select": [3, 5, 6], "If": [3, 5, 6, 8, 11], "moethod": [], "abov": 5, "locat": [], "execut": 3, "do": [3, 5, 7, 8, 10, 13], "zip": [], "shown": [], "fig": [], "enter": [6, 13], "setup": [5, 7], "red": [], "ellips": [], "thi": [3, 5, 6, 9, 10, 11, 12, 13], "restart": 3, "readi": 3, "now": [3, 6, 7, 9], "should": [3, 5, 6, 7, 13], "an": [4, 5, 6, 7, 8, 10, 11, 12], "how": 4, "get": [4, 5, 7, 8, 9, 12], "variou": [4, 5, 9], "them": [4, 5, 7, 13], "The": [5, 6, 7, 11, 13], "contain": [5, 6, 8, 13], "assort": 5, "routin": [5, 13], "been": [5, 11, 13], "packag": 5, "For": 5, "detail": 5, "inform": [5, 6, 8], "about": 5, "function": [5, 6, 7, 8, 11, 13], "check": [5, 6], "document": [5, 6], "normal": [5, 6, 10], "exampl": [5, 6, 7, 8, 10, 11, 12, 13], "b": [5, 7, 13], "point": [5, 7, 13], "without": 5, "error": [5, 7, 8, 9, 11, 13], "bar": 5, "through": [5, 13], "set": [5, 6, 7, 8, 9, 12, 13], "spline": [5, 13], "curv": [5, 8, 13], "like": [5, 6], "access": [5, 6, 11], "variabl": [5, 6, 7, 11, 12, 13], "directli": [5, 6, 13], "via": 5, "name": [5, 6, 9, 11, 12, 13], "straight": 5, "polynomi": [5, 8], "gener": [5, 7, 8, 10, 12], "linear": [5, 7, 8, 13], "non": [5, 7], "kwarg": [5, 6, 7, 12, 13], "open": [5, 6, 11, 12, 13], "read": [5, 6, 10, 11], "md": 5, "d": [5, 6, 11], "var": [5, 9], "valu": [5, 6, 7, 9, 11, 12, 13], "my_var": [5, 11, 12], "arrai": [5, 6, 7, 8, 11, 13], "mv": 5, "old": [5, 7], "better": 5, "calibr": 5, "none": [5, 6, 7, 8, 9, 10, 11, 12, 13], "spectrum": [5, 10], "convert": [5, 6, 11, 12], "h": 5, "co60": 5, "spe": [5, 10], "class": [5, 6, 7, 8, 9, 10, 11, 12], "rang": 5, "bin": 5, "bin_error": 5, "bin_cent": 5, "bin_cont": 5, "titl": [5, 13], "my": [5, 6, 11], "xlabel": 5, "x": [5, 6, 7, 8, 11, 12, 13], "ylabel": 5, "defin": [5, 6, 7, 8, 9, 13], "np": [5, 6, 13], "wai": [5, 6], "one": [5, 6, 7, 11], "1d": 5, "hi": 5, "output": 5, "bc": 5, "center": 5, "bcont": 5, "filenam": [5, 6, 11, 12], "store": [3, 5, 6, 8, 9, 11, 12, 13], "given": [5, 6, 11], "usual": 5, "result": [5, 6, 7, 8, 11, 13], "save": [5, 6, 7, 11, 12], "keyword": [5, 6, 7, 9, 11, 13], "mean": [5, 7, 8, 9, 13], "lower": [5, 7], "upper": [5, 7], "limit": [5, 6], "e": [5, 6, 7, 12, 13], "g": [5, 7, 12, 13], "10": [5, 12], "20": 5, "number": [5, 6, 7, 13], "each": [5, 6, 9, 11], "load": [5, 6, 7], "zoom": 5, "label": [5, 13], "y": [5, 6, 7, 8, 13], "addit": [5, 6, 7, 8, 13], "argument": [5, 6, 7], "pass": [5, 7, 13], "apply_calibr": 5, "cal": 5, "appli": [5, 6], "axi": [5, 13], "new": [5, 6, 7, 11, 12], "xaxi": 5, "clear": 5, "0": [5, 6, 7, 8, 9, 12, 13], "clear_window": 5, "reset": 5, "copi": [5, 7], "creat": [5, 6, 8, 10, 11, 12], "hc": 5, "labl": 5, "etc": 5, "fill": 5, "add": [5, 6, 12], "fals": [5, 6, 8, 11, 12, 13], "true": [5, 6, 7, 8, 11, 12, 13], "exist": [5, 6], "find_bin": 5, "find": [5, 6, 8], "would": [5, 6], "xmin": 5, "xmax": 5, "init": 5, "ignore_zero": 5, "gaussian": 5, "quadrat": 5, "background": 5, "can": [5, 6, 11, 12], "also": [5, 6, 8, 11, 13], "just": [5, 12, 13], "control": 5, "which": [5, 6, 7, 11, 13], "anoth": [5, 11], "part": 5, "provid": [5, 6, 8, 11, 13], "reason": [5, 7], "guess": 5, "object": [5, 6, 8, 9, 11, 12, 13], "own": 5, "properti": 5, "see": [5, 12, 13], "full": [5, 12], "follow": [5, 6, 8, 9, 11, 12], "f": [5, 6, 7, 8], "b_0": 5, "b_1x": 5, "b_2x": 5, "2": [5, 6, 7, 8, 12, 13], "exp": [5, 7], "mu": 5, "sigma": 5, "member": [5, 8], "b_o": 5, "b0": 5, "b_1": 5, "b1": 5, "b_2": 5, "b2": 5, "set_fit_list": 5, "default": [5, 6, 7, 13], "estim": 5, "initi": [5, 7], "automat": [5, 6, 7], "ignor": [5, 6], "channel": [5, 10], "zero": 5, "know": 5, "what": [5, 6], "fit_func": 5, "fit_view": 5, "current": [5, 6, 7, 9], "displai": 5, "ha": [5, 6, 7, 11], "init_gauss": 5, "calcul": [5, 13], "call": [5, 6, 7, 8, 9, 11], "init_paramet": 5, "return": [5, 6, 7, 9, 11], "doe": [3, 5, 6], "ymin": 5, "ax": [5, 8, 13], "draw": 5, "horizont": 5, "plot_fit": [5, 7, 8], "color": [5, 13], "r": [5, 8], "rebin": 5, "n": [5, 8], "scale": [5, 6, 13], "use_mean": 5, "replac": 5, "factor": [5, 6, 13], "origin": 5, "multipl": 5, "last": 5, "version": [5, 7], "format": [5, 6, 11, 12], "list": [5, 6, 7, 8, 11, 13], "set_window": 5, "similar": [5, 6, 13], "region": 5, "interest": 5, "roi": 5, "set_window_view": 5, "show_fit_list": 5, "show": [5, 7, 8, 13], "sum": 5, "between": 5, "within": 5, "when": [5, 6, 12, 13], "specifi": [5, 6], "s0": 5, "s1": 5, "5": [5, 6, 7, 8, 12, 13], "s2": 5, "x_valu": 5, "y_valu": 5, "x_bin_cent": 5, "y_bin_cent": 5, "zlabel": 5, "bad_color": 5, "w": [5, 6, 11], "colorbar": 5, "logz": 5, "2d": [5, 8], "histogram2d": 5, "xv": [5, 7, 13], "yx": 5, "h2": 5, "yv": [5, 13], "his2": 5, "xbc": 5, "ybc": 5, "xvalu": 5, "yvalu": 5, "binsx": 5, "binsi": 5, "z": [5, 6], "below": [5, 7, 12, 13], "zmin": 5, "log": [5, 13], "cal_x": 5, "cal_i": 5, "yaxi": 5, "vaku": 5, "draw_polygon": 5, "p": [5, 7, 8, 9, 13], "polygon": 5, "cut": 5, "coordin": 5, "pair": [5, 11], "determin": [5, 7], "corner": 5, "graph": 5, "patch": 5, "clevel": 5, "colormap": 5, "linearsegmentedcolormap": 5, "skip_x_label": 5, "skip_ylabel": 5, "skip_titl": 5, "type": [5, 6, 7, 11, 13], "contour": 5, "surfac": 5, "lego": 5, "level": 5, "cmrmap": 5, "logarithm": 5, "skip_y_label": 5, "poss": 5, "poly_cut": 5, "triangl": 5, "4": [5, 6, 8, 12, 13], "3": [5, 6, 7, 12, 13], "mask": 5, "project_x": 5, "project": 5, "onto": 5, "includ": [5, 6, 13], "slice": 5, "both": [5, 6, 11], "project_i": 5, "rect_cut": 5, "x1": 5, "x2": 5, "y1": 5, "y2": 5, "rectangl": 5, "sel": 5, "write": [5, 6, 11], "set_nan": 5, "err_valu": 5, "nan": 5, "correspond": [5, 6], "insid": 5, "poli": [5, 8], "rect": 5, "In": [3, 5, 11, 12], "hr": 5, "h1": 5, "subtract": 5, "mutipl": 5, "combin": [5, 12], "note": [5, 6, 13], "ad": [5, 6], "thei": [5, 6, 9, 13], "same": 5, "column": [6, 11], "dictionari": [6, 8, 11], "accord": [6, 11], "header": [6, 12], "comment": [6, 11, 12], "first": [6, 11], "It": 6, "must": [6, 12], "preced": 6, "repres": 6, "dtype": [6, 11], "col": 6, "nr": 6, "option": [6, 7, 8], "separ": [6, 11], "white": 6, "space": 6, "NOT": 6, "comma": 6, "": [6, 9], "string": [6, 9, 11, 12], "float": [6, 11, 12], "integ": 6, "blank": 6, "p_miss": [6, 12], "siglt": [6, 12], "s01": [6, 12], "alt": [6, 12], "200": [6, 12], "35e": [6, 12], "220": [6, 12], "56e": [6, 12], "230": [6, 12], "47e": [6, 12], "6": [6, 12], "f0": [6, 8], "sig_lt": 6, "dat": 6, "loop": 6, "over": 6, "l": 6, "pm": 6, "10000": 6, "print": [6, 8, 9, 11], "end": 6, "attribut": [6, 11], "make_attr": [6, 11], "kei": [6, 11, 13], "csv": 6, "export": 6, "than": [6, 8], "latex": 6, "write_csv": 6, "debug": [6, 12], "skip": 6, "use_numpi": 6, "fast": 6, "adata": 6, "interpret": [6, 12], "df": [6, 13], "my_datafil": [6, 11, 12], "iform": 6, "loadtxt": 6, "some": 6, "entri": 6, "best": 6, "larg": 6, "pure": 6, "numer": 6, "overridden": 6, "my_lin": 6, "syntax": 6, "add_data": 6, "xval": [6, 7], "yval": 6, "zval": 6, "describ": 6, "complet": 6, "add_header_com": 6, "text": 6, "begin": [6, 12], "add_kei": 6, "assign": [6, 7, 9], "newkei": 6, "some_new_valu": 6, "fp": [6, 11], "new_fil": [6, 11], "write_al": [6, 11, 12], "add_paramet": [6, 12], "section": 6, "check_data": 6, "func": [6, 8, 13], "arg": [6, 13], "fulfil": 6, "condit": 6, "user": [6, 8], "delete_kei": 6, "remov": 6, "associ": [6, 11], "eval_data": 6, "eval_str": 6, "iter": 6, "evalu": [6, 7, 8], "express": 6, "get_all_data": [6, 11], "sel_func": 6, "sel_arg": 6, "subject": 6, "selector": 6, "record": 6, "those": 6, "certain": 6, "substr": 6, "jo": 6, "def": [6, 7], "myfind": 6, "str": 6, "get_data_ev": 6, "under": 6, "get_data_list": 6, "keylist": [6, 11], "key1": [6, 11], "key2": [6, 11], "key3": [6, 11], "j": [6, 12], "charact": 6, "get_data_list_ev": 6, "select_data_ev": 6, "get_full_head": 6, "up": [6, 7], "get_head": 6, "get_kei": 6, "instanc": [6, 11], "With": 6, "written": 6, "multipli": 6, "select_data": 6, "As": [6, 11], "its": 6, "suppli": 6, "sub": 6, "eval": 6, "dataset": 6, "item": 6, "larger": [6, 8], "equal": 6, "show_all_data": [6, 11, 12], "show_data": 6, "show_kei": 6, "sort": 6, "update_head": 6, "updat": 6, "sync": 6, "complete_head": [6, 12], "write_complete_head": 6, "entir": 6, "intern": 6, "handl": [6, 7], "mydil": 6, "write_comlete_head": 6, "write_head": 6, "write_lin": 6, "write_select": 6, "index_list": 6, "ident": 6, "messag": [6, 11], "marquardt": 7, "algorithm": 7, "To": [7, 12], "term": 7, "carri": 7, "out": 7, "depend": 7, "outcom": 7, "chang": [7, 13], "decid": 7, "fix": 7, "amp": 7, "omega": 7, "intial": 7, "c": 7, "phase": 7, "sin": [7, 8], "y_exp": 7, "yerr": [7, 8], "sigi": [7, 13], "take": 7, "account": 7, "kept": 7, "constant": 7, "repeat": 7, "final": 7, "errorbar": [7, 13], "xpl": 7, "ypl": 7, "y_err": 7, "nplot": 7, "100": 7, "full_output": 7, "ftol": 7, "001": 7, "print_result": 7, "mandatori": 7, "independ": 7, "minim": 7, "chi": [7, 8], "squar": [7, 8], "converg": 7, "optim": 7, "least_squar": 7, "bound": 7, "loss": 7, "altern": [7, 11, 13], "outlier": 7, "huber": 7, "f_scale": 7, "residu": 7, "more": [7, 11, 13], "affect": 7, "leastsq": 7, "get_paramet": 7, "standard": 7, "agument": 7, "save_paramet": 7, "deep": 7, "set_xval": 7, "set_yv": 7, "show_paramet": [7, 8], "least": 8, "regress": 8, "quiet": 8, "simpl": 8, "re": 8, "offset": 8, "slope": 8, "sigma_o": 8, "sigma_": 8, "chi_r": 8, "reduc": 8, "cl": 8, "probabl": 8, "ideal": 8, "50": [8, 9], "cov": 8, "covari": 8, "matrix": 8, "order": 8, "np_scale": 8, "perform": [8, 13], "predefin": 8, "c0": 8, "const": 8, "c1": 8, "lin": 8, "c2": 8, "quad": 8, "par": [8, 11, 12], "sig_par": 8, "varianc": 8, "covar": 8, "matric": 8, "a0": 8, "a1": 8, "f1": 8, "f2": 8, "lambda": 8, "a2x": 8, "a2": 8, "a4x": 8, "sig_i": [8, 13], "err": 9, "mypar": 9, "my_first_paramet": 9, "possibl": [9, 11], "tupl": 9, "represent": 9, "ascii": [10, 11], "multi": 10, "analyz": 10, "extens": 10, "ortec": 10, "program": 10, "maestro": 10, "m": 10, "sp": 10, "usag": 10, "wa": [10, 12], "semi": 11, "colon": 11, "name1": 11, "value1": 11, "name2": 11, "value2": 11, "pf": [11, 12], "my_fil": [11, 12], "sourc": 11, "my_data": 11, "retriev": 11, "get_valu": [11, 12], "var_typ": [11, 12], "int": [11, 12], "bool": [11, 12], "agrument": 11, "tri": 11, "fail": 11, "get_valy": 11, "cannot": 11, "get_variable_nam": 11, "show_variable_nam": 11, "pointer": 11, "obtain": 11, "statment": 11, "befor": [12, 13], "noth": 12, "my_other_var": 12, "regular": 12, "pp": 12, "raw": 12, "attach": 12, "pdata": 12, "scan": 12, "file_o": 12, "set_paramet": 12, "so": [12, 13], "shortcut": 13, "typic": 13, "task": 13, "pl": 13, "pylab": 13, "dy": 13, "dyt": 13, "marker": 13, "linestyl": 13, "_nolegend_": 13, "ecolor_1": 13, "black": 13, "elinewidth": 13, "capsiz": 13, "mew": 13, "linewidth": 13, "logi": 13, "logx": 13, "min_val": 13, "x_label": 13, "y_label": 13, "plot_titl": 13, "skip_label": 13, "few": 13, "errobar": 13, "word": 13, "sig_y_tot": 13, "two": 13, "total": 13, "xerr": 13, "sig_x": 13, "style": 13, "legend": 13, "min": 13, "alli": 13, "errror": 13, "put": 13, "usus": 13, "familiar": 13, "here": 13, "along": 13, "convx": 13, "convi": 13, "mostli": 13, "5e": 13, "12": 13, "nstep": 13, "conv": 13, "interpol": 13, "increas": 13, "sigyt": 13, "implement": 13, "priviledg": 3, "work": 3, "correct": 3, "look": 3}, "objects": {"LT": [[10, 0, 0, "-", "MCA"], [5, 0, 0, "-", "box"], [6, 0, 0, "-", "datafile"], [11, 0, 0, "-", "parameterfile"], [12, 0, 0, "-", "pdatafile"], [13, 0, 0, "-", "plotting"]], "LT.MCA": [[10, 1, 1, "", "MCA"]], "LT.box": [[5, 2, 1, "", "get_data"], [5, 2, 1, "", "get_file"], [5, 2, 1, "", "get_spectrum"], [5, 1, 1, "", "histo"], [5, 1, 1, "", "histo2d"]], "LT.box.histo": [[5, 3, 1, "", "apply_calibration"], [5, 3, 1, "", "clear"], [5, 3, 1, "", "clear_window"], [5, 3, 1, "", "copy"], [5, 3, 1, "", "fill"], [5, 3, 1, "", "find_bin"], [5, 3, 1, "", "fit"], [5, 3, 1, "", "fit_func"], [5, 3, 1, "", "fit_view"], [5, 3, 1, "", "init_gauss"], [5, 3, 1, "", "init_parameters"], [5, 3, 1, "", "load"], [5, 3, 1, "", "plot"], [5, 3, 1, "", "plot_exp"], [5, 3, 1, "", "plot_fit"], [5, 3, 1, "", "rebin"], [5, 3, 1, "", "save"], [5, 3, 1, "", "set_fit_list"], [5, 3, 1, "", "set_window"], [5, 3, 1, "", "set_window_view"], [5, 3, 1, "", "show_fit_list"], [5, 3, 1, "", "sum"]], "LT.box.histo2d": [[5, 3, 1, "", "apply_calibration"], [5, 3, 1, "", "clear"], [5, 3, 1, "", "draw_polygon"], [5, 3, 1, "", "fill"], [5, 3, 1, "", "find_bin"], [5, 3, 1, "", "plot"], [5, 3, 1, "", "poly_cut"], [5, 3, 1, "", "project_x"], [5, 3, 1, "", "project_y"], [5, 3, 1, "", "rect_cut"], [5, 3, 1, "", "save"], [5, 3, 1, "", "set_nans"], [5, 3, 1, "", "sum"]], "LT.datafile": [[6, 1, 1, "", "dfile"]], "LT.datafile.dfile": [[6, 3, 1, "", "add_data"], [6, 3, 1, "", "add_header_comment"], [6, 3, 1, "", "add_key"], [6, 3, 1, "", "add_parameter"], [6, 3, 1, "", "check_data"], [6, 3, 1, "", "delete_key"], [6, 3, 1, "", "eval_data"], [6, 3, 1, "", "get_all_data"], [6, 3, 1, "", "get_data"], [6, 3, 1, "", "get_data_eval"], [6, 3, 1, "", "get_data_list"], [6, 3, 1, "", "get_data_list_eval"], [6, 3, 1, "", "get_full_header"], [6, 3, 1, "", "get_header"], [6, 3, 1, "", "get_keys"], [6, 3, 1, "", "name"], [6, 3, 1, "", "save"], [6, 3, 1, "", "scale"], [6, 3, 1, "", "select_data"], [6, 3, 1, "", "select_data_eval"], [6, 3, 1, "", "show_all_data"], [6, 3, 1, "", "show_data"], [6, 3, 1, "", "show_keys"], [6, 3, 1, "", "sort"], [6, 3, 1, "", "update_header"], [6, 3, 1, "", "write_all"], [6, 3, 1, "", "write_complete_header"], [6, 3, 1, "", "write_csv"], [6, 3, 1, "", "write_header"], [6, 3, 1, "", "write_line"], [6, 3, 1, "", "write_selected"]], "LT.parameterfile": [[11, 1, 1, "", "pfile"]], "LT.parameterfile.pfile": [[11, 3, 1, "", "Bool"], [11, 3, 1, "", "get_all_data"], [11, 3, 1, "", "get_data"], [11, 3, 1, "", "get_value"], [11, 3, 1, "", "get_variable_names"], [11, 3, 1, "", "make_attr"], [11, 3, 1, "", "name"], [11, 3, 1, "", "show_all_data"], [11, 3, 1, "", "show_variable_names"], [11, 3, 1, "", "write_all"]], "LT.pdatafile": [[12, 1, 1, "", "pdfile"]], "LT.pdatafile.pdfile": [[12, 3, 1, "", "add_parameter"], [12, 3, 1, "", "set_parameter"]], "LT.plotting": [[13, 2, 1, "", "datafile_plot_exp"], [13, 2, 1, "", "datafile_plot_theory"], [13, 2, 1, "", "datafile_spline_plot_theory"], [13, 2, 1, "", "dplot_exp"], [13, 2, 1, "", "dplot_line"], [13, 2, 1, "", "dplot_spline"], [13, 2, 1, "", "plot_exp"], [13, 2, 1, "", "plot_line"], [13, 2, 1, "", "plot_spline"]], "LT_Fit": [[7, 0, 0, "-", "gen_fit"], [8, 0, 0, "-", "linear_fit"], [9, 0, 0, "-", "parameters"]], "LT_Fit.gen_fit": [[7, 1, 1, "", "genfit"]], "LT_Fit.gen_fit.genfit": [[7, 3, 1, "", "get_parameters"], [7, 3, 1, "", "plot"], [7, 3, 1, "", "save_parameters"], [7, 3, 1, "", "set_xval"], [7, 3, 1, "", "set_yval"], [7, 3, 1, "", "show_parameters"]], "LT_Fit.linear_fit": [[8, 1, 1, "", "gen_linfit"], [8, 1, 1, "", "linefit"], [8, 1, 1, "", "polyfit"]], "LT_Fit.parameters": [[9, 1, 1, "", "Parameter"]], "LT_Fit.parameters.Parameter": [[9, 3, 1, "", "get"], [9, 3, 1, "", "set"]]}, "objtypes": {"0": "py:module", "1": "py:class", "2": "py:function", "3": "py:method"}, "objnames": {"0": ["py", "module", "Python module"], "1": ["py", "class", "Python class"], "2": ["py", "function", "Python function"], "3": ["py", "method", "Python method"]}, "titleterms": {"The": [0, 1, 2], "lt": 0, "packag": [0, 1, 2], "lt_fit": 1, "main": 2, "indic": 2, "tabl": 2, "instal": 3, "us": 3, "anaconda": 3, "introduct": 4, "tutori": 4, "box": 5, "histogram": 5, "oper": 5, "datafil": 6, "gen_fit": 7, "linear_fit": 8, "paramet": 9, "mca": 10, "parameterfil": 11, "pdatafil": 12, "plot": 13}, "envversion": {"sphinx.domains.c": 2, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 8, "sphinx.domains.index": 1, "sphinx.domains.javascript": 2, "sphinx.domains.math": 2, "sphinx.domains.python": 3, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx.ext.intersphinx": 1, "sphinx": 57}, "alltitles": {"The LT Package": [[0, "the-lt-package"]], "The LT_Fit Package": [[1, "the-lt-fit-package"]], "The Main Packages": [[2, "the-main-packages"]], "Indices and tables": [[2, "indices-and-tables"]], "Introduction/Tutorial": [[4, "introduction-tutorial"]], "Installation": [[3, "installation"]], "Installation using Anaconda": [[3, "installation-using-anaconda"]], "box": [[5, "box"]], "Histogram Operations": [[5, "histogram-operations"]], "datafile": [[6, "datafile"]], "gen_fit": [[7, "gen-fit"]], "linear_fit": [[8, "linear-fit"]], "parameters": [[9, "parameters"]], "MCA": [[10, "mca"]], "parameterfile": [[11, "parameterfile"]], "pdatafile": [[12, "pdatafile"]], "plotting": [[13, "plotting"]]}, "indexentries": {"lt.box": [[5, "module-LT.box"]], "apply_calibration() (lt.box.histo method)": [[5, "LT.box.histo.apply_calibration"]], "apply_calibration() (lt.box.histo2d method)": [[5, "LT.box.histo2d.apply_calibration"]], "clear() (lt.box.histo method)": [[5, "LT.box.histo.clear"]], "clear() (lt.box.histo2d method)": [[5, "LT.box.histo2d.clear"]], "clear_window() (lt.box.histo method)": [[5, "LT.box.histo.clear_window"]], "copy() (lt.box.histo method)": [[5, "LT.box.histo.copy"]], "draw_polygon() (lt.box.histo2d method)": [[5, "LT.box.histo2d.draw_polygon"]], "fill() (lt.box.histo method)": [[5, "LT.box.histo.fill"]], "fill() (lt.box.histo2d method)": [[5, "LT.box.histo2d.fill"]], "find_bin() (lt.box.histo method)": [[5, "LT.box.histo.find_bin"]], "find_bin() (lt.box.histo2d method)": [[5, "LT.box.histo2d.find_bin"]], "fit() (lt.box.histo method)": [[5, "LT.box.histo.fit"]], "fit_func() (lt.box.histo method)": [[5, "LT.box.histo.fit_func"]], "fit_view() (lt.box.histo method)": [[5, "LT.box.histo.fit_view"]], "get_data() (in module lt.box)": [[5, "LT.box.get_data"]], "get_file() (in module lt.box)": [[5, "LT.box.get_file"]], "get_spectrum() (in module lt.box)": [[5, "LT.box.get_spectrum"]], "histo (class in lt.box)": [[5, "LT.box.histo"]], "histo2d (class in lt.box)": [[5, "LT.box.histo2d"]], "init_gauss() (lt.box.histo method)": [[5, "LT.box.histo.init_gauss"]], "init_parameters() (lt.box.histo method)": [[5, "LT.box.histo.init_parameters"]], "load() (lt.box.histo method)": [[5, "LT.box.histo.load"]], "module": [[5, "module-LT.box"], [6, "module-LT.datafile"], [7, "module-LT_Fit.gen_fit"], [8, "module-LT_Fit.linear_fit"], [9, "module-LT_Fit.parameters"], [10, "module-LT.MCA"], [11, "module-LT.parameterfile"], [12, "module-LT.pdatafile"], [13, "module-LT.plotting"]], "plot() (lt.box.histo method)": [[5, "LT.box.histo.plot"]], "plot() (lt.box.histo2d method)": [[5, "LT.box.histo2d.plot"]], "plot_exp() (lt.box.histo method)": [[5, "LT.box.histo.plot_exp"]], "plot_fit() (lt.box.histo method)": [[5, "LT.box.histo.plot_fit"]], "poly_cut() (lt.box.histo2d method)": [[5, "LT.box.histo2d.poly_cut"]], "project_x() (lt.box.histo2d method)": [[5, "LT.box.histo2d.project_x"]], "project_y() (lt.box.histo2d method)": [[5, "LT.box.histo2d.project_y"]], "rebin() (lt.box.histo method)": [[5, "LT.box.histo.rebin"]], "rect_cut() (lt.box.histo2d method)": [[5, "LT.box.histo2d.rect_cut"]], "save() (lt.box.histo method)": [[5, "LT.box.histo.save"]], "save() (lt.box.histo2d method)": [[5, "LT.box.histo2d.save"]], "set_fit_list() (lt.box.histo method)": [[5, "LT.box.histo.set_fit_list"]], "set_nans() (lt.box.histo2d method)": [[5, "LT.box.histo2d.set_nans"]], "set_window() (lt.box.histo method)": [[5, "LT.box.histo.set_window"]], "set_window_view() (lt.box.histo method)": [[5, "LT.box.histo.set_window_view"]], "show_fit_list() (lt.box.histo method)": [[5, "LT.box.histo.show_fit_list"]], "sum() (lt.box.histo method)": [[5, "LT.box.histo.sum"]], "sum() (lt.box.histo2d method)": [[5, "LT.box.histo2d.sum"]], "lt.datafile": [[6, "module-LT.datafile"]], "add_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.add_data"]], "add_header_comment() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.add_header_comment"]], "add_key() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.add_key"]], "add_parameter() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.add_parameter"]], "check_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.check_data"]], "delete_key() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.delete_key"]], "dfile (class in lt.datafile)": [[6, "LT.datafile.dfile"]], "eval_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.eval_data"]], "get_all_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_all_data"]], "get_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_data"]], "get_data_eval() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_data_eval"]], "get_data_list() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_data_list"]], "get_data_list_eval() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_data_list_eval"]], "get_full_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_full_header"]], "get_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_header"]], "get_keys() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_keys"]], "name() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.name"]], "save() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.save"]], "scale() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.scale"]], "select_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.select_data"]], "select_data_eval() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.select_data_eval"]], "show_all_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.show_all_data"]], "show_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.show_data"]], "show_keys() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.show_keys"]], "sort() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.sort"]], "update_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.update_header"]], "write_all() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_all"]], "write_complete_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_complete_header"]], "write_csv() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_csv"]], "write_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_header"]], "write_line() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_line"]], "write_selected() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_selected"]], "lt_fit.gen_fit": [[7, "module-LT_Fit.gen_fit"]], "genfit (class in lt_fit.gen_fit)": [[7, "LT_Fit.gen_fit.genfit"]], "get_parameters() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.get_parameters"]], "plot() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.plot"]], "save_parameters() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.save_parameters"]], "set_xval() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.set_xval"]], "set_yval() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.set_yval"]], "show_parameters() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.show_parameters"]], "lt_fit.linear_fit": [[8, "module-LT_Fit.linear_fit"]], "gen_linfit (class in lt_fit.linear_fit)": [[8, "LT_Fit.linear_fit.gen_linfit"]], "linefit (class in lt_fit.linear_fit)": [[8, "LT_Fit.linear_fit.linefit"]], "polyfit (class in lt_fit.linear_fit)": [[8, "LT_Fit.linear_fit.polyfit"]], "lt_fit.parameters": [[9, "module-LT_Fit.parameters"]], "parameter (class in lt_fit.parameters)": [[9, "LT_Fit.parameters.Parameter"]], "get() (lt_fit.parameters.parameter method)": [[9, "LT_Fit.parameters.Parameter.get"]], "set() (lt_fit.parameters.parameter method)": [[9, "LT_Fit.parameters.Parameter.set"]], "lt.mca": [[10, "module-LT.MCA"]], "mca (class in lt.mca)": [[10, "LT.MCA.MCA"]], "bool() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.Bool"]], "lt.parameterfile": [[11, "module-LT.parameterfile"]], "get_all_data() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.get_all_data"]], "get_data() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.get_data"]], "get_value() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.get_value"]], "get_variable_names() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.get_variable_names"]], "make_attr() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.make_attr"]], "name() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.name"]], "pfile (class in lt.parameterfile)": [[11, "LT.parameterfile.pfile"]], "show_all_data() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.show_all_data"]], "show_variable_names() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.show_variable_names"]], "write_all() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.write_all"]], "lt.pdatafile": [[12, "module-LT.pdatafile"]], "add_parameter() (lt.pdatafile.pdfile method)": [[12, "LT.pdatafile.pdfile.add_parameter"]], "pdfile (class in lt.pdatafile)": [[12, "LT.pdatafile.pdfile"]], "set_parameter() (lt.pdatafile.pdfile method)": [[12, "LT.pdatafile.pdfile.set_parameter"]], "lt.plotting": [[13, "module-LT.plotting"]], "datafile_plot_exp() (in module lt.plotting)": [[13, "LT.plotting.datafile_plot_exp"]], "datafile_plot_theory() (in module lt.plotting)": [[13, "LT.plotting.datafile_plot_theory"]], "datafile_spline_plot_theory() (in module lt.plotting)": [[13, "LT.plotting.datafile_spline_plot_theory"]], "dplot_exp() (in module lt.plotting)": [[13, "LT.plotting.dplot_exp"]], "dplot_line() (in module lt.plotting)": [[13, "LT.plotting.dplot_line"]], "dplot_spline() (in module lt.plotting)": [[13, "LT.plotting.dplot_spline"]], "plot_exp() (in module lt.plotting)": [[13, "LT.plotting.plot_exp"]], "plot_line() (in module lt.plotting)": [[13, "LT.plotting.plot_line"]], "plot_spline() (in module lt.plotting)": [[13, "LT.plotting.plot_spline"]]}})
+Search.setIndex({"docnames": ["LT", "LT_Fit", "index", "install", "introduction", "src/lt_box", "src/lt_datafile", "src/lt_fit_gen_fit", "src/lt_fit_linear_fit", "src/lt_fit_parameters", "src/lt_mca", "src/lt_parameterfile", "src/lt_pdatafile", "src/lt_plotting"], "filenames": ["LT.rst", "LT_Fit.rst", "index.rst", "install.rst", "introduction.rst", "src/lt_box.rst", "src/lt_datafile.rst", "src/lt_fit_gen_fit.rst", "src/lt_fit_linear_fit.rst", "src/lt_fit_parameters.rst", "src/lt_mca.rst", "src/lt_parameterfile.rst", "src/lt_pdatafile.rst", "src/lt_plotting.rst"], "titles": ["The LT Package", "The LT_Fit Package", "The Main Packages", "Installation", "Introduction/Tutorial", "box", "datafile", "gen_fit", "linear_fit", "parameters", "MCA", "parameterfile", "pdatafile", "plotting"], "terms": {"A": [0, 5], "collect": 0, "python": [0, 3, 11], "softwar": [0, 3, 10], "tool": [0, 1, 5, 7, 8, 9], "data": [0, 4, 5, 6, 7, 8, 11, 12, 13], "analysi": [0, 4, 5], "present": [0, 5, 6, 12], "experiment": [0, 5, 7, 13], "physic": [0, 13], "ar": [0, 3, 5, 6, 7, 8, 9, 11, 12, 13], "base": [0, 5, 7, 13], "numpi": [0, 5, 6, 7, 8, 13], "scipi": [0, 7], "matplotlib": [0, 5, 7, 13], "recommend": [0, 3], "interact": [0, 13], "shell": 0, "i": [0, 5, 6, 7, 8, 9, 11, 12, 13], "ipython": [0, 13], "content": [0, 1, 2, 5, 6, 12, 13], "box": [0, 2, 7], "get_fil": [0, 5], "get_data": [0, 5, 6, 11, 12, 13], "get_spectrum": [0, 5], "histo": [0, 5], "histo2d": [0, 5], "histogram": 0, "oper": [0, 9], "datafil": [0, 2, 5, 11, 12, 13], "dfile": [0, 6, 12, 13], "parameterfil": [0, 2, 12], "pfile": [0, 11, 12], "pdatafil": [0, 2, 5], "pdfile": [0, 12], "plot": [0, 2, 5, 7, 8], "plot_exp": [0, 5, 7, 13], "plot_lin": [0, 5, 7, 13], "plot_splin": [0, 5, 13], "datafile_plot_exp": [0, 5, 13], "datafile_plot_theori": [0, 5, 13], "datafile_spline_plot_theori": [0, 5, 13], "dplot_exp": [0, 5, 13], "dplot_lin": [0, 5, 13], "dplot_splin": [0, 5, 13], "mca": [0, 2, 5], "fit": [1, 5, 7, 8, 9], "labtools3": [1, 3], "paramet": [1, 2, 5, 6, 7, 8, 11, 12], "linear_fit": [1, 2], "linefit": [1, 5, 8], "polyfit": [1, 5, 8], "gen_linfit": [1, 5, 8], "gen_fit": [1, 2, 5], "genfit": [1, 5, 7], "lt": [2, 4, 5, 6, 7, 10, 11, 12, 13], "lt_fit": [2, 4, 5, 7, 8, 9], "instal": [2, 6], "us": [2, 4, 5, 6, 7, 8, 9, 10, 11, 13], "anaconda": 2, "introduct": 2, "tutori": [2, 3], "index": [2, 6], "modul": [2, 4, 5, 7, 8, 10, 11, 13], "search": 2, "page": 2, "These": [3, 5, 9], "instruct": 3, "assum": [3, 5, 6, 10, 11, 13], "you": [3, 5, 6, 7, 8, 10, 13], "have": [3, 5, 10, 11], "all": [3, 5, 6, 8, 11, 12, 13], "necessari": [3, 4, 5], "requir": 3, "run": 3, "import": [3, 5, 7, 8, 9, 10, 12, 13], "onli": [3, 5, 6, 7, 9, 13], "after": 3, "ani": [3, 11, 13], "other": [3, 6, 9, 11], "distribut": 3, "want": [3, 5, 6, 8], "There": [3, 13], "sever": 3, "method": [3, 5], "avail": [3, 6], "download": 3, "script": 3, "install_labtool": 3, "py": 3, "from": [3, 5, 11, 12, 13], "command": [3, 13], "line": [3, 5, 6, 7, 8, 11, 12, 13], "pip": 3, "make": [3, 5, 6, 7, 8], "sure": [3, 5, 6], "administr": 3, "priviledg": 3, "window": [3, 5], "maco": 3, "linux": 3, "might": [3, 7], "need": [3, 5, 6, 7, 13], "sudo": 3, "start": [3, 5, 6, 11, 12], "spyder": 3, "file": [3, 5, 6, 10, 11, 12], "explor": 3, "go": 3, "directori": 3, "folder": 3, "where": [3, 4, 5, 6, 8], "store": [3, 5, 6, 8, 9, 11, 12, 13], "figur": 3, "1": [3, 5, 6, 7, 9, 12, 13], "select": [3, 5, 6], "In": [3, 5, 11, 12], "thi": [3, 5, 6, 9, 10, 11, 12, 13], "execut": 3, "If": [3, 5, 6, 8, 11], "doe": [3, 5, 6], "work": 3, "correct": 3, "look": 3, "restart": 3, "readi": 3, "now": [3, 6, 7, 9], "should": [3, 5, 6, 7, 13], "do": [3, 5, 7, 8, 10, 13], "an": [4, 5, 6, 7, 8, 10, 11, 12], "how": 4, "get": [4, 5, 7, 8, 9, 12], "variou": [4, 5, 9], "them": [4, 5, 7, 13], "labtool": 4, "The": [5, 6, 7, 11, 13], "contain": [5, 6, 8, 13], "assort": 5, "routin": [5, 13], "been": [5, 11, 13], "packag": 5, "For": 5, "detail": 5, "inform": [5, 6, 8], "about": 5, "function": [5, 6, 7, 8, 11, 13], "check": [5, 6], "document": [5, 6], "normal": [5, 6, 7, 10], "exampl": [5, 6, 7, 8, 10, 11, 12, 13], "b": [5, 7, 13], "point": [5, 7, 13], "without": 5, "error": [5, 7, 8, 9, 11, 13], "bar": 5, "through": [5, 13], "set": [5, 6, 7, 8, 9, 12, 13], "spline": [5, 13], "curv": [5, 8, 13], "like": [5, 6], "access": [5, 6, 11], "variabl": [5, 6, 7, 11, 12, 13], "directli": [5, 6, 13], "via": 5, "name": [5, 6, 9, 11, 12, 13], "straight": 5, "polynomi": [5, 8], "gener": [5, 7, 8, 10, 12], "linear": [5, 7, 8, 13], "non": [5, 7], "kwarg": [5, 6, 7, 12, 13], "open": [5, 6, 11, 12, 13], "read": [5, 6, 10, 11], "md": 5, "d": [5, 6, 11], "var": [5, 9], "valu": [5, 6, 7, 9, 11, 12, 13], "my_var": [5, 11, 12], "arrai": [5, 6, 7, 8, 11, 13], "mv": 5, "old": [5, 7], "better": 5, "calibr": 5, "none": [5, 6, 7, 8, 9, 10, 11, 12, 13], "spectrum": [5, 10], "convert": [5, 6, 11, 12], "h": [5, 7], "co60": 5, "spe": [5, 10], "class": [5, 6, 7, 8, 9, 10, 11, 12], "rang": 5, "bin": 5, "bin_error": 5, "bin_cent": 5, "bin_cont": 5, "titl": [5, 13], "my": [5, 6, 11], "xlabel": 5, "x": [5, 6, 7, 8, 11, 12, 13], "ylabel": 5, "calc_w2": 5, "fals": [5, 6, 8, 11, 12, 13], "defin": [5, 6, 7, 8, 9, 13], "np": [5, 6, 7, 13], "wai": [5, 6], "one": [5, 6, 7, 11], "1d": 5, "hi": 5, "output": 5, "bc": 5, "center": 5, "bcont": 5, "filenam": [5, 6, 11, 12], "given": [5, 6, 11], "usual": 5, "result": [5, 6, 7, 8, 11, 13], "save": [5, 6, 7, 11, 12], "keyword": [5, 6, 7, 9, 11, 13], "mean": [5, 7, 8, 9, 13], "lower": [5, 7], "upper": [5, 7], "limit": [5, 6], "e": [5, 6, 7, 12, 13], "g": [5, 7, 12, 13], "10": [5, 12], "20": [5, 7], "number": [5, 6, 7, 13], "each": [5, 6, 7, 9, 11], "load": [5, 6, 7], "zoom": 5, "label": [5, 13], "y": [5, 6, 7, 8, 13], "true": [5, 6, 7, 8, 11, 12, 13], "calcculta": 5, "sum": 5, "squar": [5, 7, 8], "weight": 5, "calcul": [5, 7, 13], "final": [5, 7], "increment": 5, "addit": [5, 6, 7, 8, 13], "argument": [5, 6, 7], "pass": [5, 7, 13], "apply_calibr": 5, "cal": 5, "appli": [5, 6], "axi": [5, 13], "new": [5, 6, 7, 11, 12], "xaxi": 5, "bkg": 5, "return": [5, 6, 7, 9, 11], "background": 5, "float": [5, 6, 11, 12], "b_val": 5, "backgound": 5, "clear": 5, "0": [5, 6, 7, 8, 9, 12, 13], "clear_window": 5, "reset": 5, "copi": [5, 7], "creat": [5, 6, 8, 10, 11, 12], "hc": 5, "labl": 5, "etc": 5, "fill": 5, "add": [5, 6, 12], "exist": [5, 6], "find_bin": 5, "find": [5, 6, 8], "would": [5, 6], "xmin": 5, "xmax": 5, "init": 5, "ignore_zero": 5, "gaussian": 5, "quadrat": 5, "can": [5, 6, 11, 12], "also": [5, 6, 8, 11, 13], "just": [5, 12, 13], "control": 5, "which": [5, 6, 7, 11, 13], "anoth": [5, 11], "part": 5, "provid": [5, 6, 8, 11, 13], "reason": [5, 7], "guess": 5, "object": [5, 6, 8, 9, 11, 12, 13], "own": 5, "properti": 5, "see": [5, 12, 13], "full": [5, 12], "follow": [5, 6, 8, 9, 11, 12], "f": [5, 6, 7, 8], "b_0": 5, "b_1x": 5, "b_2x": 5, "2": [5, 6, 7, 8, 12, 13], "exp": [5, 7], "mu": 5, "sigma": 5, "member": [5, 8], "b_o": 5, "b0": 5, "b_1": 5, "b1": 5, "b_2": 5, "b2": 5, "set_fit_list": 5, "default": [5, 6, 7, 13], "estim": 5, "initi": [5, 7], "automat": [5, 6, 7], "ignor": [5, 6], "channel": [5, 10], "zero": 5, "know": 5, "what": [5, 6], "fit_func": 5, "fit_view": 5, "current": [5, 6, 7, 9], "displai": 5, "ha": [5, 6, 7, 11], "gauss": 5, "init_gauss": 5, "call": [5, 6, 7, 8, 9, 11], "init_paramet": 5, "ymin": 5, "ax": [5, 8, 13], "draw": 5, "horizont": 5, "plot_fit": [5, 7, 8], "color": [5, 13], "r": [5, 8], "rebin": 5, "n": [5, 8], "scale": [5, 6, 7, 13], "use_mean": 5, "replac": 5, "factor": [5, 6, 13], "origin": 5, "multipl": 5, "last": [5, 11], "version": [5, 7], "format": [5, 6, 11, 12], "list": [5, 6, 7, 8, 11, 13], "set_window": 5, "similar": [5, 6, 13], "region": 5, "interest": 5, "roi": 5, "set_window_view": 5, "show_fit_list": 5, "show": [5, 7, 8, 13], "between": 5, "within": 5, "when": [5, 6, 12, 13], "specifi": [5, 6], "s0": 5, "s1": 5, "5": [5, 6, 7, 8, 12, 13], "s2": 5, "x_valu": 5, "y_valu": 5, "x_bin_cent": 5, "y_bin_cent": 5, "zlabel": 5, "bad_color": 5, "w": [5, 6, 11], "colorbar": 5, "logz": 5, "2d": [5, 8], "histogram2d": 5, "xv": [5, 7, 13], "yx": 5, "h2": 5, "yv": [5, 13], "his2": 5, "xbc": 5, "ybc": 5, "xvalu": 5, "yvalu": 5, "binsx": 5, "binsi": 5, "z": [5, 6], "below": [5, 7, 12, 13], "zmin": 5, "log": [5, 13], "cal_x": 5, "cal_i": 5, "yaxi": 5, "vaku": 5, "draw_polygon": 5, "p": [5, 7, 8, 9, 13], "polygon": 5, "cut": 5, "coordin": 5, "pair": [5, 11], "determin": [5, 7], "corner": 5, "graph": 5, "patch": 5, "clevel": 5, "colormap": 5, "linearsegmentedcolormap": 5, "skip_x_label": 5, "skip_ylabel": 5, "skip_titl": 5, "type": [5, 6, 7, 11, 13], "contour": 5, "surfac": 5, "lego": 5, "level": 5, "cmrmap": 5, "logarithm": 5, "skip_y_label": 5, "poss": 5, "poly_cut": 5, "setup": [5, 7], "triangl": 5, "4": [5, 6, 8, 12, 13], "3": [5, 6, 7, 12, 13], "mask": 5, "project_x": 5, "project": 5, "onto": 5, "includ": [5, 6, 11, 13], "slice": 5, "both": [5, 6, 11], "project_i": 5, "rect_cut": 5, "x1": 5, "x2": 5, "y1": 5, "y2": 5, "rectangl": 5, "sel": 5, "write": [5, 6, 11], "set_nan": 5, "err_valu": 5, "nan": 5, "correspond": [5, 6, 7], "insid": 5, "poli": [5, 8], "rect": 5, "hr": 5, "h1": 5, "subtract": 5, "mutipl": 5, "combin": [5, 12], "abov": 5, "note": [5, 6, 13], "ad": [5, 6, 11], "thei": [5, 6, 9, 13], "same": 5, "column": [6, 11], "dictionari": [6, 8, 11], "accord": [6, 11], "header": [6, 12], "comment": [6, 11, 12], "first": [6, 11], "It": 6, "must": [6, 12], "preced": 6, "repres": 6, "dtype": [6, 11], "col": 6, "nr": 6, "option": [6, 7, 8], "separ": [6, 11], "white": 6, "space": 6, "NOT": 6, "comma": [6, 11], "": [6, 9], "string": [6, 9, 11, 12], "integ": 6, "blank": 6, "enter": [6, 13], "p_miss": [6, 12], "siglt": [6, 12], "s01": [6, 12], "alt": [6, 12], "200": [6, 12], "35e": [6, 12], "220": [6, 12], "56e": [6, 12], "230": [6, 12], "47e": [6, 12], "6": [6, 12], "f0": [6, 8], "sig_lt": 6, "dat": 6, "loop": 6, "over": 6, "l": 6, "pm": 6, "10000": 6, "print": [6, 8, 9, 11], "end": 6, "attribut": [6, 11], "make_attr": [6, 11], "kei": [6, 11, 13], "csv": 6, "export": 6, "than": [6, 8], "latex": 6, "write_csv": 6, "debug": [6, 12], "skip": 6, "use_numpi": 6, "fast": 6, "adata": 6, "interpret": [6, 12], "df": [6, 13], "my_datafil": [6, 11, 12], "iform": 6, "loadtxt": 6, "some": 6, "entri": 6, "best": 6, "larg": 6, "pure": 6, "numer": [6, 7], "overridden": 6, "my_lin": 6, "syntax": 6, "add_data": 6, "xval": [6, 7], "yval": 6, "zval": 6, "describ": 6, "complet": 6, "add_header_com": 6, "text": 6, "begin": [6, 12], "add_kei": 6, "assign": [6, 7, 9], "newkei": 6, "some_new_valu": 6, "fp": [6, 11], "new_fil": [6, 11], "write_al": [6, 11, 12], "add_paramet": [6, 12], "section": 6, "check_data": 6, "func": [6, 8, 13], "arg": [6, 13], "fulfil": 6, "condit": 6, "user": [6, 8], "delete_kei": 6, "remov": [6, 11], "associ": [6, 11], "eval_data": 6, "eval_str": 6, "iter": 6, "evalu": [6, 7, 8], "express": 6, "get_all_data": [6, 11], "sel_func": 6, "sel_arg": 6, "subject": 6, "selector": 6, "record": 6, "those": 6, "certain": 6, "substr": 6, "jo": 6, "def": [6, 7], "myfind": 6, "str": 6, "get_data_ev": 6, "under": 6, "get_data_list": 6, "keylist": [6, 11], "key1": [6, 11], "key2": [6, 11], "key3": [6, 11], "j": [6, 12], "charact": [6, 11], "get_data_list_ev": 6, "select_data_ev": 6, "get_full_head": 6, "up": [6, 7], "get_head": 6, "get_kei": 6, "instanc": [6, 11], "With": 6, "written": 6, "multipli": 6, "select_data": 6, "As": [6, 11], "its": 6, "suppli": 6, "sub": 6, "eval": 6, "dataset": 6, "item": 6, "larger": [6, 8], "equal": 6, "show_all_data": [6, 11, 12], "show_data": 6, "show_kei": 6, "sort": 6, "update_head": 6, "updat": 6, "sync": 6, "complete_head": [6, 12], "write_complete_head": 6, "entir": 6, "intern": 6, "handl": [6, 7], "mydil": 6, "write_comlete_head": 6, "write_head": 6, "write_lin": 6, "write_select": 6, "index_list": 6, "ident": 6, "messag": [6, 11], "marquardt": 7, "algorithm": 7, "To": [7, 12], "term": 7, "carri": 7, "out": 7, "depend": 7, "outcom": 7, "chang": [7, 13], "decid": 7, "fix": 7, "amp": 7, "omega": 7, "intial": 7, "c": 7, "phase": 7, "sin": [7, 8], "pseudo": 7, "linspac": 7, "y_exp": 7, "random": 7, "loc": 7, "05": 7, "size": 7, "len": 7, "sig_i": [7, 8, 13], "ones_lik": 7, "y_err": 7, "take": 7, "account": 7, "kept": 7, "constant": 7, "repeat": 7, "deriv": 7, "analyt": 7, "dfda": 7, "lambda": [7, 8], "wrsp": 7, "dfdb": 7, "co": 7, "dfdc": 7, "deriv_list": 7, "func_deriv": 7, "errorbar": [7, 13], "xpl": 7, "ypl": 7, "nplot": 7, "100": 7, "full_output": 7, "ftol": 7, "001": 7, "diff_step": 7, "print_result": 7, "mandatori": 7, "independ": 7, "minim": 7, "chi": [7, 8], "converg": 7, "step": 7, "derivet": 7, "respect": 7, "par": [7, 8, 11, 12], "optim": 7, "least_squar": 7, "bound": 7, "loss": 7, "altern": [7, 11, 13], "outlier": 7, "huber": 7, "f_scale": 7, "residu": 7, "more": [7, 11, 13], "affect": 7, "leastsq": 7, "get_paramet": 7, "standard": 7, "agument": 7, "save_paramet": 7, "deep": 7, "set_xval": 7, "set_yv": 7, "show_paramet": [7, 8], "least": 8, "regress": 8, "yerr": 8, "quiet": 8, "simpl": 8, "re": 8, "offset": 8, "slope": 8, "sigma_o": 8, "sigma_": 8, "chi_r": 8, "reduc": 8, "cl": 8, "probabl": 8, "ideal": 8, "50": [8, 9], "cov": 8, "covari": 8, "matrix": 8, "order": 8, "np_scale": 8, "perform": [8, 13], "predefin": 8, "c0": 8, "const": 8, "c1": 8, "lin": 8, "c2": 8, "quad": 8, "sig_par": 8, "varianc": 8, "covar": 8, "matric": 8, "a0": 8, "a1": 8, "f1": 8, "f2": 8, "a2x": 8, "a2": 8, "a4x": 8, "err": 9, "mypar": 9, "my_first_paramet": 9, "possibl": [9, 11], "tupl": 9, "represent": 9, "ascii": [10, 11], "multi": 10, "analyz": 10, "extens": 10, "ortec": 10, "program": 10, "maestro": 10, "m": 10, "sp": 10, "usag": 10, "wa": [10, 12], "semi": 11, "colon": 11, "name1": 11, "value1": 11, "name2": 11, "value2": 11, "pf": [11, 12], "my_fil": [11, 12], "sourc": 11, "my_data": 11, "retriev": 11, "get_valu": [11, 12], "var_typ": [11, 12], "int": [11, 12], "bool": [11, 12], "agrument": 11, "tri": 11, "fail": 11, "input": 11, "continu": 11, "concaten": 11, "backslash": 11, "while": 11, "no_file_provid": 11, "get_valy": 11, "cannot": 11, "get_variable_nam": 11, "show_variable_nam": 11, "pointer": 11, "obtain": 11, "statment": 11, "befor": [12, 13], "noth": 12, "my_other_var": 12, "regular": 12, "pp": 12, "raw": 12, "attach": 12, "pdata": 12, "scan": 12, "file_o": 12, "set_paramet": 12, "so": [12, 13], "shortcut": 13, "typic": 13, "task": 13, "pl": 13, "pylab": 13, "dy": 13, "dyt": 13, "marker": 13, "linestyl": 13, "_nolegend_": 13, "ecolor_1": 13, "black": 13, "elinewidth": 13, "capsiz": 13, "mew": 13, "linewidth": 13, "logi": 13, "logx": 13, "min_val": 13, "x_label": 13, "y_label": 13, "plot_titl": 13, "skip_label": 13, "few": 13, "errobar": 13, "word": 13, "sig_y_tot": 13, "two": 13, "total": 13, "xerr": 13, "sig_x": 13, "style": 13, "legend": 13, "min": 13, "alli": 13, "errror": 13, "put": 13, "usus": 13, "familiar": 13, "here": 13, "along": 13, "convx": 13, "convi": 13, "mostli": 13, "5e": 13, "12": 13, "nstep": 13, "conv": 13, "interpol": 13, "increas": 13, "sigi": 13, "sigyt": 13, "implement": 13}, "objects": {"LT": [[10, 0, 0, "-", "MCA"], [5, 0, 0, "-", "box"], [6, 0, 0, "-", "datafile"], [11, 0, 0, "-", "parameterfile"], [12, 0, 0, "-", "pdatafile"], [13, 0, 0, "-", "plotting"]], "LT.MCA": [[10, 1, 1, "", "MCA"]], "LT.box": [[5, 2, 1, "", "get_data"], [5, 2, 1, "", "get_file"], [5, 2, 1, "", "get_spectrum"], [5, 1, 1, "", "histo"], [5, 1, 1, "", "histo2d"]], "LT.box.histo": [[5, 3, 1, "", "apply_calibration"], [5, 3, 1, "", "bkg"], [5, 3, 1, "", "clear"], [5, 3, 1, "", "clear_window"], [5, 3, 1, "", "copy"], [5, 3, 1, "", "fill"], [5, 3, 1, "", "find_bin"], [5, 3, 1, "", "fit"], [5, 3, 1, "", "fit_func"], [5, 3, 1, "", "fit_view"], [5, 3, 1, "", "gauss"], [5, 3, 1, "", "init_gauss"], [5, 3, 1, "", "init_parameters"], [5, 3, 1, "", "load"], [5, 3, 1, "", "plot"], [5, 3, 1, "", "plot_exp"], [5, 3, 1, "", "plot_fit"], [5, 3, 1, "", "rebin"], [5, 3, 1, "", "save"], [5, 3, 1, "", "set_fit_list"], [5, 3, 1, "", "set_window"], [5, 3, 1, "", "set_window_view"], [5, 3, 1, "", "show_fit_list"], [5, 3, 1, "", "sum"]], "LT.box.histo2d": [[5, 3, 1, "", "apply_calibration"], [5, 3, 1, "", "clear"], [5, 3, 1, "", "draw_polygon"], [5, 3, 1, "", "fill"], [5, 3, 1, "", "find_bin"], [5, 3, 1, "", "plot"], [5, 3, 1, "", "poly_cut"], [5, 3, 1, "", "project_x"], [5, 3, 1, "", "project_y"], [5, 3, 1, "", "rect_cut"], [5, 3, 1, "", "save"], [5, 3, 1, "", "set_nans"], [5, 3, 1, "", "sum"]], "LT.datafile": [[6, 1, 1, "", "dfile"]], "LT.datafile.dfile": [[6, 3, 1, "", "add_data"], [6, 3, 1, "", "add_header_comment"], [6, 3, 1, "", "add_key"], [6, 3, 1, "", "add_parameter"], [6, 3, 1, "", "check_data"], [6, 3, 1, "", "delete_key"], [6, 3, 1, "", "eval_data"], [6, 3, 1, "", "get_all_data"], [6, 3, 1, "", "get_data"], [6, 3, 1, "", "get_data_eval"], [6, 3, 1, "", "get_data_list"], [6, 3, 1, "", "get_data_list_eval"], [6, 3, 1, "", "get_full_header"], [6, 3, 1, "", "get_header"], [6, 3, 1, "", "get_keys"], [6, 3, 1, "", "name"], [6, 3, 1, "", "save"], [6, 3, 1, "", "scale"], [6, 3, 1, "", "select_data"], [6, 3, 1, "", "select_data_eval"], [6, 3, 1, "", "show_all_data"], [6, 3, 1, "", "show_data"], [6, 3, 1, "", "show_keys"], [6, 3, 1, "", "sort"], [6, 3, 1, "", "update_header"], [6, 3, 1, "", "write_all"], [6, 3, 1, "", "write_complete_header"], [6, 3, 1, "", "write_csv"], [6, 3, 1, "", "write_header"], [6, 3, 1, "", "write_line"], [6, 3, 1, "", "write_selected"]], "LT.parameterfile": [[11, 1, 1, "", "pfile"]], "LT.parameterfile.pfile": [[11, 3, 1, "", "Bool"], [11, 3, 1, "", "get_all_data"], [11, 3, 1, "", "get_data"], [11, 3, 1, "", "get_value"], [11, 3, 1, "", "get_variable_names"], [11, 3, 1, "", "make_attr"], [11, 3, 1, "", "name"], [11, 3, 1, "", "show_all_data"], [11, 3, 1, "", "show_variable_names"], [11, 3, 1, "", "write_all"]], "LT.pdatafile": [[12, 1, 1, "", "pdfile"]], "LT.pdatafile.pdfile": [[12, 3, 1, "", "add_parameter"], [12, 3, 1, "", "set_parameter"]], "LT.plotting": [[13, 2, 1, "", "datafile_plot_exp"], [13, 2, 1, "", "datafile_plot_theory"], [13, 2, 1, "", "datafile_spline_plot_theory"], [13, 2, 1, "", "dplot_exp"], [13, 2, 1, "", "dplot_line"], [13, 2, 1, "", "dplot_spline"], [13, 2, 1, "", "plot_exp"], [13, 2, 1, "", "plot_line"], [13, 2, 1, "", "plot_spline"]], "LT_Fit": [[7, 0, 0, "-", "gen_fit"], [8, 0, 0, "-", "linear_fit"], [9, 0, 0, "-", "parameters"]], "LT_Fit.gen_fit": [[7, 1, 1, "", "genfit"]], "LT_Fit.gen_fit.genfit": [[7, 3, 1, "", "get_parameters"], [7, 3, 1, "", "plot"], [7, 3, 1, "", "save_parameters"], [7, 3, 1, "", "set_xval"], [7, 3, 1, "", "set_yval"], [7, 3, 1, "", "show_parameters"]], "LT_Fit.linear_fit": [[8, 1, 1, "", "gen_linfit"], [8, 1, 1, "", "linefit"], [8, 1, 1, "", "polyfit"]], "LT_Fit.parameters": [[9, 1, 1, "", "Parameter"]], "LT_Fit.parameters.Parameter": [[9, 3, 1, "", "get"], [9, 3, 1, "", "set"]]}, "objtypes": {"0": "py:module", "1": "py:class", "2": "py:function", "3": "py:method"}, "objnames": {"0": ["py", "module", "Python module"], "1": ["py", "class", "Python class"], "2": ["py", "function", "Python function"], "3": ["py", "method", "Python method"]}, "titleterms": {"The": [0, 1, 2], "lt": 0, "packag": [0, 1, 2], "lt_fit": 1, "main": 2, "indic": 2, "tabl": 2, "instal": 3, "us": 3, "anaconda": 3, "introduct": 4, "tutori": 4, "box": 5, "histogram": 5, "oper": 5, "datafil": 6, "gen_fit": 7, "linear_fit": 8, "paramet": 9, "mca": 10, "parameterfil": 11, "pdatafil": 12, "plot": 13}, "envversion": {"sphinx.domains.c": 2, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 8, "sphinx.domains.index": 1, "sphinx.domains.javascript": 2, "sphinx.domains.math": 2, "sphinx.domains.python": 3, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx.ext.intersphinx": 1, "sphinx": 57}, "alltitles": {"The LT Package": [[0, "the-lt-package"]], "The LT_Fit Package": [[1, "the-lt-fit-package"]], "The Main Packages": [[2, "the-main-packages"]], "Indices and tables": [[2, "indices-and-tables"]], "Installation": [[3, "installation"]], "Installation using Anaconda": [[3, "installation-using-anaconda"]], "Introduction/Tutorial": [[4, "introduction-tutorial"]], "box": [[5, "box"]], "Histogram Operations": [[5, "histogram-operations"]], "datafile": [[6, "datafile"]], "gen_fit": [[7, "gen-fit"]], "linear_fit": [[8, "linear-fit"]], "parameters": [[9, "parameters"]], "MCA": [[10, "mca"]], "parameterfile": [[11, "parameterfile"]], "pdatafile": [[12, "pdatafile"]], "plotting": [[13, "plotting"]]}, "indexentries": {"lt.box": [[5, "module-LT.box"]], "apply_calibration() (lt.box.histo method)": [[5, "LT.box.histo.apply_calibration"]], "apply_calibration() (lt.box.histo2d method)": [[5, "LT.box.histo2d.apply_calibration"]], "bkg() (lt.box.histo method)": [[5, "LT.box.histo.bkg"]], "clear() (lt.box.histo method)": [[5, "LT.box.histo.clear"]], "clear() (lt.box.histo2d method)": [[5, "LT.box.histo2d.clear"]], "clear_window() (lt.box.histo method)": [[5, "LT.box.histo.clear_window"]], "copy() (lt.box.histo method)": [[5, "LT.box.histo.copy"]], "draw_polygon() (lt.box.histo2d method)": [[5, "LT.box.histo2d.draw_polygon"]], "fill() (lt.box.histo method)": [[5, "LT.box.histo.fill"]], "fill() (lt.box.histo2d method)": [[5, "LT.box.histo2d.fill"]], "find_bin() (lt.box.histo method)": [[5, "LT.box.histo.find_bin"]], "find_bin() (lt.box.histo2d method)": [[5, "LT.box.histo2d.find_bin"]], "fit() (lt.box.histo method)": [[5, "LT.box.histo.fit"]], "fit_func() (lt.box.histo method)": [[5, "LT.box.histo.fit_func"]], "fit_view() (lt.box.histo method)": [[5, "LT.box.histo.fit_view"]], "gauss() (lt.box.histo method)": [[5, "LT.box.histo.gauss"]], "get_data() (in module lt.box)": [[5, "LT.box.get_data"]], "get_file() (in module lt.box)": [[5, "LT.box.get_file"]], "get_spectrum() (in module lt.box)": [[5, "LT.box.get_spectrum"]], "histo (class in lt.box)": [[5, "LT.box.histo"]], "histo2d (class in lt.box)": [[5, "LT.box.histo2d"]], "init_gauss() (lt.box.histo method)": [[5, "LT.box.histo.init_gauss"]], "init_parameters() (lt.box.histo method)": [[5, "LT.box.histo.init_parameters"]], "load() (lt.box.histo method)": [[5, "LT.box.histo.load"]], "module": [[5, "module-LT.box"], [6, "module-LT.datafile"], [7, "module-LT_Fit.gen_fit"], [8, "module-LT_Fit.linear_fit"], [9, "module-LT_Fit.parameters"], [10, "module-LT.MCA"], [11, "module-LT.parameterfile"], [12, "module-LT.pdatafile"], [13, "module-LT.plotting"]], "plot() (lt.box.histo method)": [[5, "LT.box.histo.plot"]], "plot() (lt.box.histo2d method)": [[5, "LT.box.histo2d.plot"]], "plot_exp() (lt.box.histo method)": [[5, "LT.box.histo.plot_exp"]], "plot_fit() (lt.box.histo method)": [[5, "LT.box.histo.plot_fit"]], "poly_cut() (lt.box.histo2d method)": [[5, "LT.box.histo2d.poly_cut"]], "project_x() (lt.box.histo2d method)": [[5, "LT.box.histo2d.project_x"]], "project_y() (lt.box.histo2d method)": [[5, "LT.box.histo2d.project_y"]], "rebin() (lt.box.histo method)": [[5, "LT.box.histo.rebin"]], "rect_cut() (lt.box.histo2d method)": [[5, "LT.box.histo2d.rect_cut"]], "save() (lt.box.histo method)": [[5, "LT.box.histo.save"]], "save() (lt.box.histo2d method)": [[5, "LT.box.histo2d.save"]], "set_fit_list() (lt.box.histo method)": [[5, "LT.box.histo.set_fit_list"]], "set_nans() (lt.box.histo2d method)": [[5, "LT.box.histo2d.set_nans"]], "set_window() (lt.box.histo method)": [[5, "LT.box.histo.set_window"]], "set_window_view() (lt.box.histo method)": [[5, "LT.box.histo.set_window_view"]], "show_fit_list() (lt.box.histo method)": [[5, "LT.box.histo.show_fit_list"]], "sum() (lt.box.histo method)": [[5, "LT.box.histo.sum"]], "sum() (lt.box.histo2d method)": [[5, "LT.box.histo2d.sum"]], "lt.datafile": [[6, "module-LT.datafile"]], "add_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.add_data"]], "add_header_comment() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.add_header_comment"]], "add_key() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.add_key"]], "add_parameter() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.add_parameter"]], "check_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.check_data"]], "delete_key() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.delete_key"]], "dfile (class in lt.datafile)": [[6, "LT.datafile.dfile"]], "eval_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.eval_data"]], "get_all_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_all_data"]], "get_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_data"]], "get_data_eval() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_data_eval"]], "get_data_list() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_data_list"]], "get_data_list_eval() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_data_list_eval"]], "get_full_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_full_header"]], "get_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_header"]], "get_keys() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.get_keys"]], "name() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.name"]], "save() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.save"]], "scale() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.scale"]], "select_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.select_data"]], "select_data_eval() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.select_data_eval"]], "show_all_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.show_all_data"]], "show_data() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.show_data"]], "show_keys() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.show_keys"]], "sort() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.sort"]], "update_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.update_header"]], "write_all() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_all"]], "write_complete_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_complete_header"]], "write_csv() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_csv"]], "write_header() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_header"]], "write_line() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_line"]], "write_selected() (lt.datafile.dfile method)": [[6, "LT.datafile.dfile.write_selected"]], "lt_fit.gen_fit": [[7, "module-LT_Fit.gen_fit"]], "genfit (class in lt_fit.gen_fit)": [[7, "LT_Fit.gen_fit.genfit"]], "get_parameters() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.get_parameters"]], "plot() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.plot"]], "save_parameters() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.save_parameters"]], "set_xval() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.set_xval"]], "set_yval() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.set_yval"]], "show_parameters() (lt_fit.gen_fit.genfit method)": [[7, "LT_Fit.gen_fit.genfit.show_parameters"]], "lt_fit.linear_fit": [[8, "module-LT_Fit.linear_fit"]], "gen_linfit (class in lt_fit.linear_fit)": [[8, "LT_Fit.linear_fit.gen_linfit"]], "linefit (class in lt_fit.linear_fit)": [[8, "LT_Fit.linear_fit.linefit"]], "polyfit (class in lt_fit.linear_fit)": [[8, "LT_Fit.linear_fit.polyfit"]], "lt_fit.parameters": [[9, "module-LT_Fit.parameters"]], "parameter (class in lt_fit.parameters)": [[9, "LT_Fit.parameters.Parameter"]], "get() (lt_fit.parameters.parameter method)": [[9, "LT_Fit.parameters.Parameter.get"]], "set() (lt_fit.parameters.parameter method)": [[9, "LT_Fit.parameters.Parameter.set"]], "lt.mca": [[10, "module-LT.MCA"]], "mca (class in lt.mca)": [[10, "LT.MCA.MCA"]], "bool() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.Bool"]], "lt.parameterfile": [[11, "module-LT.parameterfile"]], "get_all_data() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.get_all_data"]], "get_data() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.get_data"]], "get_value() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.get_value"]], "get_variable_names() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.get_variable_names"]], "make_attr() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.make_attr"]], "name() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.name"]], "pfile (class in lt.parameterfile)": [[11, "LT.parameterfile.pfile"]], "show_all_data() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.show_all_data"]], "show_variable_names() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.show_variable_names"]], "write_all() (lt.parameterfile.pfile method)": [[11, "LT.parameterfile.pfile.write_all"]], "lt.pdatafile": [[12, "module-LT.pdatafile"]], "add_parameter() (lt.pdatafile.pdfile method)": [[12, "LT.pdatafile.pdfile.add_parameter"]], "pdfile (class in lt.pdatafile)": [[12, "LT.pdatafile.pdfile"]], "set_parameter() (lt.pdatafile.pdfile method)": [[12, "LT.pdatafile.pdfile.set_parameter"]], "lt.plotting": [[13, "module-LT.plotting"]], "datafile_plot_exp() (in module lt.plotting)": [[13, "LT.plotting.datafile_plot_exp"]], "datafile_plot_theory() (in module lt.plotting)": [[13, "LT.plotting.datafile_plot_theory"]], "datafile_spline_plot_theory() (in module lt.plotting)": [[13, "LT.plotting.datafile_spline_plot_theory"]], "dplot_exp() (in module lt.plotting)": [[13, "LT.plotting.dplot_exp"]], "dplot_line() (in module lt.plotting)": [[13, "LT.plotting.dplot_line"]], "dplot_spline() (in module lt.plotting)": [[13, "LT.plotting.dplot_spline"]], "plot_exp() (in module lt.plotting)": [[13, "LT.plotting.plot_exp"]], "plot_line() (in module lt.plotting)": [[13, "LT.plotting.plot_line"]], "plot_spline() (in module lt.plotting)": [[13, "LT.plotting.plot_spline"]]}})
